* pmap_rmt.c, svc.c, svc_simple.c, svc_udp.c, svc_tcp.c,         xdr_array.c, xdr_reference.c : Include string.h for memset and         memcmp prototypes.
        * auth_gssapi.c, auth_gssapi_misc.c, svc_auth_gssapi.c: Cast
        argment to PRINTF to match format.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13595 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/rpc/auth_gssapi.c
@@ -394,7 +394,7 @@ next_token:
 		    goto cleanup;
 	       } else if (isn_buf.length != sizeof(rpc_u_int32)) {
 		    PRINTF(("gssapi_create: gss_unseal gave %d bytes\n",
-			    isn_buf.length));
+			    (int) isn_buf.length));
 		    goto cleanup;
 	       }
